Convert toString utils to TS
