--- packages/rrule-generator/src/utils/computeRRule/toString/computeDaily.js
@@ -1,8 +1,0 @@
-import RRule from 'rrule';
-
-const computeDaily = ({ interval }) => ({
-  freq: RRule.DAILY,
-  interval,
-});
-
-export default computeDaily;
