+++ packages/rrule-generator/src/utils/computeRRule/toString/computeDaily.ts
@@ -1,0 +1,9 @@
+import RRule, { Options } from 'rrule';
+import { RRuleState } from '../../../state';
+
+const computeDaily = ({ interval }: RRuleState['repeat']['daily']): Partial<Options> => ({
+	freq: RRule.DAILY,
+	interval,
+});
+
+export default computeDaily;
