--- packages/rrule-generator/src/utils/computeRRule/toString/computeEnd.js
@@ -1,17 +1,0 @@
-import moment from 'moment';
-
-const computeEnd = ({ mode, after, date }) => {
-  const end = {};
-
-  if (mode === 'After') {
-    end.count = after;
-  }
-
-  if (mode === 'On date') {
-    end.until = moment(date).format();
-  }
-
-  return end;
-};
-
-export default computeEnd;
