--- packages/rrule-generator/src/utils/computeRRule/toString/computeHourly.js
@@ -1,8 +1,0 @@
-import RRule from 'rrule';
-
-const computeHourly = ({ interval }) => ({
-  freq: RRule.HOURLY,
-  interval,
-});
-
-export default computeHourly;
