--- packages/rrule-generator/src/utils/computeRRule/toString/computeMonthlyOn.js
@@ -1,5 +1,0 @@
-const computeMonthlyOn = on => ({
-  bymonthday: on.day,
-});
-
-export default computeMonthlyOn;
