+++ packages/rrule-generator/src/utils/computeRRule/toString/computeMonthlyOn.ts
@@ -1,0 +1,8 @@
+import { Options } from 'rrule';
+import { RRuleState } from '../../../state';
+
+const computeMonthlyOn = (on: RRuleState['repeat']['monthly']['on']): Partial<Options> => ({
+	bymonthday: on.day as number,
+});
+
+export default computeMonthlyOn;
