--- packages/rrule-generator/src/utils/computeRRule/toString/computeOptions.js
@@ -1,17 +1,0 @@
-import RRule from 'rrule';
-
-const computeOptions = ({ hideStart, weekStartsOnSunday }) => {
-  const options = {};
-
-  if (hideStart) {
-    options.dtstart = null;
-  }
-
-  if (weekStartsOnSunday) {
-    options.wkst = RRule.SU;
-  }
-
-  return options;
-};
-
-export default computeOptions;
