+++ packages/rrule-generator/src/utils/computeRRule/toString/computeRRule.ts
@@ -4,8 +4,12 @@ import computeStart from './computeStart';
 import computeRepeat from './computeRepeat';
 import computeEnd from './computeEnd';
 import computeOptions from './computeOptions';
+import { RRuleState } from '../../../state';
+import { RRuleConfig } from '../../../types';
 
-const computeRRule = ({ start, repeat, end, options }) => {
+type ComputeRRule = (state: RRuleState, config: RRuleConfig, hideStart?: boolean) => string;
+
+const computeRRule: ComputeRRule = ({ start, repeat, end }, { weekStartsOn }, hideStart) => {
 	const rruleObject = {
 		...computeStart(start),
 		...computeRepeat(repeat),
