--- packages/rrule-generator/src/utils/computeRRule/toString/computeStart.js
@@ -1,16 +1,0 @@
-import moment from 'moment';
-
-const computeStart = ({ date }) => {
-  // verify that incoming date is valid
-  // by seeing if it can be converted into a moment object.
-  // if not, then create a new date
-  if (!moment.isMoment(moment(date))) {
-    date = new Date().setMilliseconds(0);
-  }
-
-  return {
-    dtstart: moment(date).toDate(),
-  };
-};
-
-export default computeStart;
