--- packages/rrule-generator/src/utils/computeRRule/toString/computeWeekly.js
@@ -1,14 +1,0 @@
-import RRule from 'rrule';
-import { values } from 'lodash';
-
-const computeWeekly = ({ interval, days }) => ({
-  freq: RRule.WEEKLY,
-  interval,
-  byweekday: values(days).reduce(
-    (activeDays, isDayActive, dayIndex) =>
-      (isDayActive ? [...activeDays, dayIndex] : activeDays),
-    [],
-  ),
-});
-
-export default computeWeekly;
