+++ packages/rrule-generator/src/utils/computeRRule/toString/computeYearly.ts
@@ -1,0 +1,13 @@
+import RRule, { Options } from 'rrule';
+
+import { RRuleState } from '../../../state';
+
+import computeYearlyOn from './computeYearlyOn';
+import computeYearlyOnThe from './computeYearlyOnThe';
+
+const computeYearly = ({ mode, on, onThe }: RRuleState['repeat']['yearly']): Partial<Options> => ({
+	freq: RRule.YEARLY,
+	...(mode === 'ON' ? computeYearlyOn(on) : computeYearlyOnThe(onThe)),
+});
+
+export default computeYearly;
