--- packages/rrule-generator/src/utils/computeRRule/toString/computeYearlyOn.js
@@ -1,8 +1,0 @@
-import { MONTHS } from '../../../constants/index';
-
-const computeYearlyOn = on => ({
-  bymonth: MONTHS.indexOf(on.month) + 1,
-  bymonthday: on.day,
-});
-
-export default computeYearlyOn;
