+++ packages/rrule-generator/src/utils/computeRRule/toString/computeYearlyOn.ts
@@ -1,0 +1,11 @@
+import { Options } from 'rrule';
+
+import { MONTHS } from '../../../constants/index';
+import { RRuleState } from '../../../state';
+
+const computeYearlyOn = (on: RRuleState['repeat']['yearly']['on']): Partial<Options> => ({
+	bymonth: MONTHS.indexOf(on.month) + 1,
+	bymonthday: on.day as number,
+});
+
+export default computeYearlyOn;
